Moved screenshot option to TABTestCase
